[JENKINS-31648] Introduce TopLevelItem abstraction.
